价格字段修改

+++ app/product/biz/model/product.go
@@ -7,13 +7,13 @@ import (
 
 type Product struct {
 	Base
-	Name          string  `gorm:"column:name"`
-	Description   string  `gorm:"column:description"`
-	Price         int64   `gorm:"column:price"`
-	OriginalPrice float32 `gorm:"column:original_price"`
-	Images        string  `gorm:"column:images"`
-	Stock         uint32  `gorm:"column:stock"`
-	Status        int     `gorm:"column:status"`
+	Name          string `gorm:"column:name"`
+	Description   string `gorm:"column:description"`
+	Price         int64  `gorm:"column:price"`
+	OriginalPrice int64  `gorm:"column:original_price"`
+	Images        string `gorm:"column:images"`
+	Stock         uint32 `gorm:"column:stock"`
+	Status        int    `gorm:"column:status"`
 }
 
 func (p Product) TableName() string {
